avoid -Wdouble-promotion warning

+++ test.c
@@ -62,7 +62,7 @@ main(void) {
 	printf("mask %" PRIX32 "\n", pcg32_rand_fast(rng32, 1UL << 20));
 	printf("fast %" PRIX32 "\n", pcg32_rand_fast(rng32, INT32_MAX));
 	printf("slow %" PRIX32 "\n", pcg32_rand(rng32, INT32_MAX));
-	printf("fp %.8f\n", pcg32_float(rng32));
+	printf("fp %.8f\n", (double)pcg32_float(rng32));
 
 	pcg64_t rng64[] = { pcg64_getentropy() };
 	assert(pcg64_rand(rng64, 0) == 0);
